refactor: Move to the top of the post component components
게시물 컴포넌트를 Movies의 하위 폴더가 아닌 components의 최상단으로 이동시켰다.

close #85
