+++ src/components/ModalContent/modalContent.scss
@@ -1,5 +1,5 @@
-@import 'src/styles/constants/_colors.scss';
-@import 'src/styles/base/_fonts.scss';
+@import '../../styles/constants/_colors.scss';
+@import '../../styles/base/_fonts.scss';
 
 .modal-contents {
   display: flex;
